feat: Remove unused files and update API routes
This commit removes unused files, including `CheckOutView.vue` and `counter.js`, which are no longer needed in the project. It also updates the API routes in `urls.py` to include new endpoints for `PedidoViewSet`, `DetallePagoViewSet`, and `ClienteViewSet`. These changes clean up the codebase and ensure that the API routes are up to date with the latest functionality.